move GapicCodeGeneratorAnnotationsTest and its test resources to a separate package
