+++ .travis.yml
@@ -9,7 +9,7 @@ script:
   - ./gradlew check --info
   - ./testutils/check-baselines.sh
   - go test testutils/check-calling-form-in-baseline/*.go
-  - go run $(find testutils/check-calling-form-in-baseline -type f | grep -v _test) -yaml src/test/java/com/google/api/codegen/testsrc/libraryproto/config_not_annotated/testdata/library_gapic.yaml src/test
+  - go run $(find testutils/check-calling-form-in-baseline -type f | grep -v _test) -yaml src/test/java/com/google/api/codegen/testsrc/libraryproto/library_gapic.yaml src/test
   - go test testutils/check-lang-uses-form/*.go
   - go run $(find testutils/check-lang-uses-form -type f | grep -v _test) -forms src/main/java/com/google/api/codegen/viewmodel/CallingForm.java src/test
 after_success:
