+++ src/test/java/com/google/api/codegen/advising/AdviserTest.java
@@ -45,9 +45,7 @@ public class AdviserTest extends ConfigBaselineTestCase {
   @Before
   public void setup() {
     getTestDataLocator().addTestDataSource(CodegenTestUtil.class, "testsrc/common");
-    getTestDataLocator()
-        .addTestDataSource(
-            CodegenTestUtil.class, "testsrc/libraryproto/config_not_annotated/testdata");
+    getTestDataLocator().addTestDataSource(CodegenTestUtil.class, "testsrc/libraryproto");
   }
 
   @Test
