+++ src/test/java/com/google/api/codegen/config/GapicConfigProducerTest.java
@@ -37,8 +37,7 @@ public class GapicConfigProducerTest {
   public void missingConfigSchemaVersion() {
     TestDataLocator locator = TestDataLocator.create(GapicConfigProducerTest.class);
     locator.addTestDataSource(CodegenTestUtil.class, "testsrc/common");
-    locator.addTestDataSource(
-        CodegenTestUtil.class, "testsrc/libraryproto/config_not_annotated/testdata");
+    locator.addTestDataSource(CodegenTestUtil.class, "testsrc/libraryproto");
     model =
         CodegenTestUtil.readModel(
             locator, tempDir, new String[] {"myproto.proto"}, new String[] {"myproto.yaml"});
