+++ src/test/java/com/google/api/codegen/protoannotations/testdata/library.proto
