--- src/test/java/com/google/api/codegen/testsrc/libraryproto/annotated_no_config/README.md
@@ -1,1 +1,0 @@
-This directory contains the version of library.proto that is proto-annotated.
