--- src/test/java/com/google/api/codegen/testsrc/libraryproto/config_not_annotated/README.md
@@ -1,5 +1,0 @@
-This directory contains the version of library.proto that
-does *not* have proto-annotations for configuration.
-The directory also contains the service yaml, packaging config file, and GAPIC config file.
-
-These test resources are used to test GAPIC configgen and client-generation with configs.
