+++ src/test/java/com/google/api/codegen/testsrc/libraryproto/library.yaml
