+++ src/test/java/com/google/api/codegen/testsrc/libraryproto/library_pkg2.yaml
