+++ src/test/java/com/google/api/codegen/transformer/ModelTypeNameConverterTestUtil.java
@@ -34,8 +34,7 @@ public class ModelTypeNameConverterTestUtil {
     String fileName = "library.proto";
     TestDataLocator locator = TestDataLocator.create(CodegenTestUtil.class);
     locator.addTestDataSource(CodegenTestUtil.class, "testsrc/common");
-    locator.addTestDataSource(
-        CodegenTestUtil.class, "testsrc/libraryproto/config_not_annotated/testdata");
+    locator.addTestDataSource(CodegenTestUtil.class, "testsrc/libraryproto");
     Model model =
         CodegenTestUtil.readModel(
             locator, tempDir, new String[] {fileName}, new String[] {"library.yaml"});
